refactor: move BackgroundIntent implementation to TodayTasksWidget.swift
- Removed the obsolete BackgroundIntent.swift file.
- Streamlined the BackgroundIntent code for better organization and clarity.
- Enhanced widget update logic to ensure reliable task processing in the background.
